Fix print warnings for Alan 2 terp under 64-bit OS.

+++ trunk/terps/alan2/debug.c
@@ -50,7 +50,7 @@ static void showatrs(atradr)
 
   i = 1;
   for (at = (AtrElem *) addrTo(atradr); !endOfTable(at); at++) {
-    sprintf(str, "$i%3ld: %ld (%s)", i, at->val, (char *) addrTo(at->stradr));
+    sprintf(str, "$i%3ld: %ld (%s)", (long) i, (unsigned long) at->val, (char *) addrTo(at->stradr));
 #if ISO == 0
     fromIso(str, str);
 #endif
@@ -71,7 +71,7 @@ static void showobjs()
 
   output("OBJECTS:");
   for (obj = OBJMIN; obj <= OBJMAX; obj++) {
-    sprintf(str, "$i%3ld: ", obj);
+    sprintf(str, "$i%3ld: ", (long) obj);
     output(str);
     say(obj);
   }
@@ -92,7 +92,7 @@ static void showobj(obj)
 
 
   if (!isObj(obj)) {
-    sprintf(str, "Object number out of range. Between %ld and %ld, please.", OBJMIN, OBJMAX);
+    sprintf(str, "Object number out of range. Between %ld and %ld, please.", (unsigned long) OBJMIN, (unsigned long) OBJMAX);
     output(str);
     return;
   }
@@ -101,7 +101,7 @@ static void showobj(obj)
   output(str);
   say(obj);
 
-  sprintf(str, "$iLocation = %ld", where(obj));
+  sprintf(str, "$iLocation = %ld", (unsigned long) where(obj));
   output(str);
   if (isLoc(objs[OBJ].loc))
     say(objs[OBJ].loc);
@@ -139,7 +139,7 @@ static void showcnts()
 
   output("CONTAINERS:");
   for (cnt = CNTMIN; cnt <= CNTMAX; cnt++) {
-    sprintf(str, "$i%3ld: ", cnt);
+    sprintf(str, "$i%3ld: ", (long) cnt);
     output(str);
     if (cnts[CNT].nam != 0)
       interpret(cnts[CNT].nam);
@@ -166,7 +166,7 @@ static void showcnt(cnt)
 #define  CNT (cnt-CNTMIN)
 
   if (cnt < CNTMIN || cnt > CNTMAX) {
-    sprintf(str, "Container number out of range. Between %ld and %ld, please.", CNTMIN, CNTMAX);
+    sprintf(str, "Container number out of range. Between %ld and %ld, please.", (unsigned long) CNTMIN, (unsigned long) CNTMAX);
     output(str);
     return;
   }
@@ -178,7 +178,7 @@ static void showcnt(cnt)
   if (cnts[CNT].parent != 0) {
     cnt = cnts[CNT].parent;
     say(cnt);
-    sprintf(str, "$iLocation = %ld", where(cnt));
+    sprintf(str, "$iLocation = %ld", (unsigned long) where(cnt));
     output(str);
   }
   output("$iContains ");
@@ -211,7 +211,7 @@ static void showlocs()
 
   output("LOCATIONS:");
   for (loc = LOCMIN; loc <= LOCMAX; loc++) {
-    sprintf(str, "$i%3ld: ", loc);
+    sprintf(str, "$i%3ld: ", (long) loc);
     output(str);
     say(loc);
   }
@@ -231,7 +231,7 @@ static void showloc(loc)
 
   
   if (!isLoc(loc)) {
-    sprintf(str, "Location number out of range. Between %ld and %ld, please.", LOCMIN, LOCMAX);
+    sprintf(str, "Location number out of range. Between %ld and %ld, please.", (unsigned long) LOCMIN, (unsigned long) LOCMAX);
     output(str);
     return;
   }
@@ -256,7 +256,7 @@ static void showacts()
 
   output("ACTORS:");
   for (act = ACTMIN; act <= ACTMAX; act++) {
-    sprintf(str, "$i%3ld:", act);
+    sprintf(str, "$i%3ld:", (long) act);
     output(str);
     say(act);
   }
@@ -276,7 +276,7 @@ static void showact(act)
   Boolean oldstp;
   
   if (!isAct(act)) {
-    sprintf(str, "Actor number out of range. Between %ld and %ld, please.", ACTMIN, ACTMAX);
+    sprintf(str, "Actor number out of range. Between %ld and %ld, please.", (unsigned long) ACTMIN, (unsigned long) ACTMAX);
     output(str);
     return;
   }
@@ -287,7 +287,7 @@ static void showact(act)
   say(act);
   stpflg = oldstp;
 
-  sprintf(str, "$iLocation = %ld", acts[act-ACTMIN].loc);
+  sprintf(str, "$iLocation = %ld", (unsigned long) acts[act-ACTMIN].loc);
   output(str);
   if (isLoc(acts[act-ACTMIN].loc))
     say(acts[act-ACTMIN].loc);
@@ -296,10 +296,10 @@ static void showact(act)
   else
     output("Illegal location!");
 
-  sprintf(str, "$iScript = %ld", acts[act-ACTMIN].script);
+  sprintf(str, "$iScript = %ld", (unsigned long) acts[act-ACTMIN].script);
   output(str);
 
-  sprintf(str, "$iStep = %ld", acts[act-ACTMIN].step);
+  sprintf(str, "$iStep = %ld", (unsigned long) acts[act-ACTMIN].step);
   output(str);
 
   output("$iAttributes =");
